fix: 🐛 reset selection default options
